+++ src/monster/monster-util.h
@@ -9,6 +9,7 @@
 enum summon_type : int;
 enum class MonraceId : short;
 class PlayerType;
+class MonsterEntity;
 void get_mon_num_prep_enum(PlayerType *player_ptr, MonraceHook hook1 = MonraceHook::NONE, MonraceHookTerrain hook2 = MonraceHookTerrain::NONE);
 void get_mon_num_prep_escort(PlayerType *player_ptr, MonraceId escorted_monrace_id, short m_idx, MonraceHookTerrain hook2);
 
@@ -57,3 +58,4 @@ void get_mon_num_prep_bounty(PlayerType *player_ptr);
 void mark_monsters_present(PlayerType *player_ptr);
 bool is_player(MONSTER_IDX m_idx);
 bool is_monster(MONSTER_IDX m_idx);
+void move_monster_to(PlayerType *player_ptr, MonsterEntity &monster, const Pos2D &pos_to);
